refactor ovn events code
move the code to their own file and add unit tests.
